3 hours Spent
I started very late on this test. I have spent around 3 hours on this.
Everything is done except that there is one bug.
If a table is reserved it is updated and displayed and the drop down menu is updated as required with respect to date and hour selected. But if I reserve another table , than the previously reserved table is showed again. as of now I cannot figure out why this is happening. And I have to sleep now. It is some stupid bug that I cannot see right now.